Created server and moved assets into public folder
