+++ package.json
@@ -4,7 +4,8 @@
   "description": "UCB second boot camp project.",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest",
+    "start": "node index.js"
   },
   "repository": {
     "type": "git",
@@ -16,5 +17,8 @@
   "bugs": {
     "url": "https://github.com/Z9tails/Project_2/issues"
   },
-  "homepage": "https://github.com/Z9tails/Project_2#readme"
+  "homepage": "https://github.com/Z9tails/Project_2#readme",
+  "dependencies": {
+    "express": "^4.18.1"
+  }
 }
