+++ public/assets/js/script.js
