+++ public/index.html
@@ -74,7 +74,7 @@
     
     <script src="./assets/js/script.js"></script>
     <!-- script js file for game -->
-    <script src="index.js"></script>
+    <script src="../src/game.js" type="module"></script>
     <!-- end game script -->
 </body>
 </html>
