+++ routes/htmlRoutes.js
@@ -1,0 +1,16 @@
+// Dependencies
+const path = require('path');
+const router = require('express').Router();
+
+// This has to be on the top than the other routers
+router.get('/', (req, res) => {
+    res.sendFile(path.join(__dirname, '../public/index.html'));
+});
+
+// If no matching route is found default to home page
+router.get('*', (req, res) => {
+    res.sendFile(path.join(__dirname, '../public/index.html'));
+});
+
+
+module.exports = router;
